+++ R/pkg.R
@@ -14,3 +14,7 @@ pkg_version <- function(pkg = pkg_name()) {
 pkg_file <- function(...) {
   system.file(..., package = pkg_name())
 }
+
+is_pkg_avail <- function(pkg) {
+  requireNamespace(pkg, quietly = TRUE)
+}