+++ NEWS.md
@@ -4,6 +4,8 @@
   component `cond`.
 * Export `get_board_option_value()` to make available current option settings
   via `session$UserData`.
+* Introduce (optional) dependency on [thematic](
+  https://rstudio.github.io/thematic/) to auto-style plots.
 
 # blockr.core 0.1.0